Upgrade to latest Rust

+++ source/rust/vndf/src/game_service/updater.rs
@@ -11,7 +11,7 @@ pub fn init(frame_time_in_ms: u64, game: Sender<GameEvent>) {
 	spawn(proc() {
 		loop {
 			game.send(Update(frame_time_in_ms as f64 / 1000.0));
-			sleep(Duration::milliseconds(frame_time_in_ms as i32));
+			sleep(Duration::milliseconds(frame_time_in_ms as i64));
 		}
 	});
 }
